"total" rewrite of malloc structure: - realloc can resize - blocks >128k are handled by mmu :)

+++ include/sys/mman.h
@@ -106,7 +106,7 @@ extern int munmap (void *__addr, size_t __len) __THROW;
 extern int mprotect (void *__addr, size_t __len, int __prot) __THROW;
 extern int msync (void *__addr, size_t __len, int __flags) __THROW;
 extern void *mremap (void *__addr, size_t __old_len, size_t __new_len,
-		     int __may_move) __THROW;
+		     unsigned long __may_move) __THROW;
 extern int mincore (void *__start, size_t __len, unsigned char *__vec);
 
 #ifdef __quad_t
